interview_prep/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
notebooklm-py>=0.3.3
